+++ internal/repo/clickhouse/migrations/1_create_table_guesses.up.sql
@@ -1,12 +1,12 @@
 CREATE TABLE IF NOT EXISTS guesses
 (
-    `user_id` int,
+    `user_id` Int64,
     `game_id` UUID,
-    `country_id` int,
+    `country_id` UInt8,
     `text` String,
-    `guess_number` int,
+    `guess_number` UInt8,
     `right` boolean,
     `timestamp` DateTime
 )
-    ENGINE = MergeTree
-    ORDER BY (user_id, timestamp);
+ENGINE = MergeTree
+ORDER BY (user_id, timestamp);
